CineMachine Camera Added
cinemachine camera 2d camera added. Also cinemachine conffiner2d added to camera. in this way it stops when it reach the boundries (we determined the boundries with collider (cameraboundreis object in scene)).
I create a new layer for boundries  to avoid trigger with onether objects like bullets and I disabled collision of this layers with other layer (edit -> projectsettings -> physic2d -> layer collisionmatrix)